Introduce a macro for debugging calls to setContextClassLoader
And some API additions: (parent cl), (dynamic-classloader thread), (cl-id cl)

Stop the priority-classloader from throwing an error when .getResource is called
with an absolute path. This is due to (io/file base file) not liking an absolute
file, and makes sense, classpath resolution should be relative to the classpath.
We do print a warning because something inside cider or nrepl is doing this and
I want to know what it is.